requirements.txt
Django==4.2.23
djangorestframework==3.16.0
django-cors-headers==4.7.0
